requirements.txt
flask>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
beautifulsoup4>=4.12.0
ShopifyAPI
python-dotenv
